+++ lib/papierkram_api/v1/endpoints/banking/transactions.rb
@@ -5,13 +5,25 @@ module PapierkramApi
     module Endpoints
       module Banking
         # This class is responsible for all the API calls related to banking transactions.
-        class Transactions
+        class Transactions < PapierkramApi::V1::Endpoints::Base
           def by(id:)
-            # wip
+            get("#{@url_api_path}/banking/transactions/#{id}")
           end
 
-          def all
-            # wip
+          def all(bank_connection_id:,
+                  page: 1,
+                  page_size: 100,
+                  order_by: nil,
+                  order_direction: nil)
+            query = {
+              bank_connection_id: bank_connection_id,
+              page: page,
+              page_size: page_size
+            }
+            query[:order_by] = order_by if order_by
+            query[:order_direction] = order_direction if order_direction
+
+            get("#{@url_api_path}/banking/transactions", query)
           end
         end
       end